Add files via upload
Introduction:
Coronavirus disease (COVID-19) is an infectious disease caused by a newly discovered coronavirus.
Most people infected with the COVID-19 virus will experience mild to moderate respiratory illness and recover without requiring special treatment.  Older people, and those with underlying medical problems like cardiovascular disease, diabetes, chronic respiratory disease, and cancer are more likely to develop serious illness.

The best way to prevent and slow down transmission is to be well informed about the COVID-19 virus, the disease it causes and how it spreads. Protect yourself and others from infection by washing your hands or using an alcohol based rub frequently and not touching your face. 

Aim:
This project shows the situation of the covid-19 curve on a daily basis and to remain relevant and re-usable everyday, I have used a web scraping method from the worldometers.info website. Hence the data used in this notebook does not have to be updated everyday; instead, the dataset is crated right here in this notebook from the above mentioned website. It uses a diverse range of data visualization graphs including stacked bar charts, pie charts and so on.